CLM-29931 - Fix severity float conversion issue while saving coordinate security row (#11236)

+++ insight-brain-data/src/test/java/com/sonatype/insight/brain/dataaccess/thirdpartyscans/ThirdPartyCoordinateSecurityDAOTest.java
@@ -165,7 +165,7 @@ public class ThirdPartyCoordinateSecurityDAOTest
       final String refId,
       final String description,
       final String link,
-      final float score,
+      final double score,
       final String fixedBy,
       final String cooedinateFileId, final ThirdPartyCoordinateSecurity actual)
   {
